Refactor estimateMessageFee function
