+++ src/starknet-wrappers.ts
@@ -2,12 +2,16 @@ import { Image, ProcessResult } from "@nomiclabs/hardhat-docker";
 import { PLUGIN_NAME, StarknetChainId } from "./constants";
 import { StarknetDockerProxy } from "./starknet-docker-proxy";
 import { StarknetVenvProxy } from "./starknet-venv-proxy";
-import { BlockNumber, InteractChoice } from "./types";
-import { adaptUrl } from "./utils";
+import { BlockNumber, InteractChoice, Numeric } from "./types";
+import { adaptUrl, numericToHexString } from "./utils";
 import { getPrefixedCommand, normalizeVenvPath } from "./utils/venv";
 import { ExternalServer } from "./external-server";
 import { StarknetPluginError } from "./starknet-plugin-error";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
+import { FeeEstimation } from "./starknet-types";
+import { hash } from "starknet";
+import { hexToDecimalString } from "starknet/utils/number";
+import axios from "axios";
 
 interface CompileWrapperOptions {
     file: string;
@@ -393,6 +397,33 @@ export abstract class StarknetWrapper {
             throw new StarknetPluginError(executed.stderr.toString());
         }
         return executed;
+    }
+
+    public async estimateMessageFee(
+        fromAddress: string,
+        toAddress: string,
+        functionName: string,
+        payload: Numeric[]
+    ): Promise<FeeEstimation> {
+        const body = {
+            from_address: hexToDecimalString(fromAddress),
+            to_address: toAddress,
+            entry_point_selector: hash.getSelectorFromName(functionName),
+            payload: payload.map((item) => numericToHexString(item))
+        };
+
+        const response = await axios.post(
+            `${this.hre.starknet.networkConfig.url}/feeder_gateway/estimate_message_fee`,
+            body
+        );
+
+        const { gas_price, gas_usage, overall_fee, unit } = response.data;
+        return {
+            amount: BigInt(overall_fee),
+            unit,
+            gas_price: BigInt(gas_price),
+            gas_usage: BigInt(gas_usage)
+        };
     }
 }
